extract videos and images into new media directory
